fix:hr user moved to worker
